+++ dash/dash.py
@@ -569,6 +569,7 @@ class Dash(ObsoleteChecker):
         self.callback_map = {}
         # same deps as a list to catch duplicate outputs, and to send to the front end
         self._callback_list = []
+        self.callback_api_paths = {}
 
         # list of inline scripts
         self._inline_scripts = []
@@ -782,6 +783,54 @@ class Dash(ObsoleteChecker):
 
         # catch-all for front-end routes, used by dcc.Location
         self._add_url("<path:path>", self.index)
+
+    def setup_apis(self):
+        """
+        Register API endpoints for all callbacks defined using `dash.callback`.
+
+        This method must be called after all callbacks are registered and before the app is served.
+        It ensures that all callback API routes are available for the Dash app to function correctly.
+
+        Typical usage:
+            app = Dash(__name__)
+            # Register callbacks here
+            app.setup_apis()
+            app.run()
+
+        If not called, callback endpoints will not be available and the app will not function as expected.
+        """
+        for k in list(_callback.GLOBAL_API_PATHS):
+            if k in self.callback_api_paths:
+                raise DuplicateCallback(
+                    f"The callback `{k}` provided with `dash.callback` was already "
+                    "assigned with `app.callback`."
+                )
+            self.callback_api_paths[k] = _callback.GLOBAL_API_PATHS.pop(k)
+
+        def make_parse_body(func):
+            def _parse_body():
+                if flask.request.is_json:
+                    data = flask.request.get_json()
+                    return flask.jsonify(func(**data))
+                return flask.jsonify({})
+
+            return _parse_body
+
+        def make_parse_body_async(func):
+            async def _parse_body_async():
+                if flask.request.is_json:
+                    data = flask.request.get_json()
+                    result = await func(**data)
+                    return flask.jsonify(result)
+                return flask.jsonify({})
+
+            return _parse_body_async
+
+        for path, func in self.callback_api_paths.items():
+            if asyncio.iscoroutinefunction(func):
+                self._add_url(path, make_parse_body_async(func), ["POST"])
+            else:
+                self._add_url(path, make_parse_body(func), ["POST"])
 
     def _setup_plotlyjs(self):
         # pylint: disable=import-outside-toplevel
@@ -1364,6 +1413,7 @@ class Dash(ObsoleteChecker):
             config_prevent_initial_callbacks=self.config.prevent_initial_callbacks,
             callback_list=self._callback_list,
             callback_map=self.callback_map,
+            callback_api_paths=self.callback_api_paths,
             **_kwargs,
         )
 
@@ -1516,6 +1566,7 @@ class Dash(ObsoleteChecker):
     def _setup_server(self):
         if self._got_first_request["setup_server"]:
             return
+
         self._got_first_request["setup_server"] = True
 
         # Apply _force_eager_loading overrides from modules